Refactor libpympb
